factor out find_completeness_tests
