+++ lib/davinci_crd_test_kit/client/v2.2.1/cross_hook/client_prefetch_complete_and_subset_test.rb
@@ -1,10 +1,12 @@
 require_relative '../../tagged_request_load_helper'
+require_relative '../../cross_hook_helper'
 require_relative '../client_urls'
 
 module DaVinciCRDTestKit
   module V221
     class ClientPrefetchCompleteAndSubsetTest < Inferno::Test
       include TaggedRequestLoadHelper
+      include CrossHookHelper
 
       title 'Client can provide both the complete standard prefetch data set and a subset'
       id :crd_v221_client_prefetch_complete_and_subset
@@ -57,29 +59,6 @@ module DaVinciCRDTestKit
 
         skip_if error_messages?,
                 'The client did not demonstrate both complete and subset prefetch capability. See Messages for details.'
-      end
-
-      def find_completeness_tests # rubocop:disable Metrics/CyclomaticComplexity
-        # minimal error checking - want it to bomb if this search logic fails
-        hooks_group = self.class.parent.parent.parent.groups.find do |group|
-          group.id.to_s.include?('crd_v221_client_hooks')
-        end
-        completeness_tests = hooks_group.groups.map do |group|
-          group.groups.flat_map(&:tests).find do |test|
-            test.id.to_s.include?('crd_v221_hook_request_prefetch_complete')
-          end
-        end
-
-        cross_hooks_interaction_group = self.class.parent.parent.groups.find do |group|
-          group.id.to_s.include?('crd_v221_client_cross_hook_interaction')
-        end
-        cross_hooks_completeness_test =
-          cross_hooks_interaction_group.groups.flat_map(&:tests).find do |test|
-            test.id.to_s.include?('crd_v221_hook_request_prefetch_complete')
-          end
-        completeness_tests << cross_hooks_completeness_test
-
-        completeness_tests.compact
       end
 
       def check_for_demonstration(requests, completeness_tests, target_name)
